ajusta navBar
